finish implementation of state persist
