+++ src/screens/MapList.tsx
@@ -1,24 +1,11 @@
 import { StatusBar } from "expo-status-bar";
-import React, { useCallback } from "react";
+import React from "react";
 import { StyleSheet, View, ActivityIndicator, Text } from "react-native";
-import { useFocusEffect } from "@react-navigation/native";
 import { LocationList } from "@components";
-import {
-  useAppDispatch,
-  getLocationsAsync,
-  useAppSelector,
-  selectLocations,
-} from "@store";
+import { useAppSelector, selectLocations } from "@store";
 
 export default function MapList() {
-  const dispatch = useAppDispatch();
   const { loading, locations, error } = useAppSelector(selectLocations);
-
-  useFocusEffect(
-    useCallback(() => {
-      dispatch(getLocationsAsync());
-    }, [dispatch, getLocationsAsync])
-  );
 
   if (loading)
     return (
