+++ src/service/soupKitchenService.ts
@@ -2,7 +2,7 @@ import { Api } from "@apis";
 import { SoupKitchenModel } from "@store";
 
 // This address is unique to your computer, please modify ipAddress accordingly. Instructions are in README.
-const url = "https://d2a9-108-50-197-203.ngrok.io";
+const url = "https://d84e-108-50-197-203.ngrok.io";
 
 interface ISoupKitchenService {
   getLocations(): Promise<SoupKitchenModel[]>;
