+++ src/store/index.ts
@@ -1,52 +1,23 @@
-import AsyncStorage from "@react-native-async-storage/async-storage";
 import {
-  persistStore,
-  persistReducer,
-  FLUSH,
-  REHYDRATE,
-  PAUSE,
-  PERSIST,
-  PURGE,
-  REGISTER,
-} from "redux-persist";
-import { combineReducers, configureStore } from "@reduxjs/toolkit";
-import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
-import soupKitchenReducer, {
   selectLocations,
   getLocationsAsync,
   SoupKitchenModel,
 } from "./soupKitchen/slice";
+import {
+  store,
+  persistor,
+  RootState,
+  useAppDispatch,
+  useAppSelector,
+} from "./store";
 
-export { selectLocations, getLocationsAsync, SoupKitchenModel };
-
-const persistConfig = {
-  key: "root",
-  version: 1,
-  storage: AsyncStorage,
-  whitelist: ["soupKitchen"],
+export {
+  selectLocations,
+  getLocationsAsync,
+  SoupKitchenModel,
+  store,
+  persistor,
+  RootState,
+  useAppDispatch,
+  useAppSelector,
 };
-
-const rootReducer = combineReducers({
-  soupKitchen: soupKitchenReducer,
-});
-
-const persistedReducer = persistReducer(persistConfig, rootReducer);
-
-// had to keep this in same file to not get warnings about require cycles...very annoying :P
-export const store = configureStore({
-  reducer: persistedReducer,
-  middleware: (getDefaultMiddleware) =>
-    getDefaultMiddleware({
-      serializableCheck: {
-        ignoredActions: [FLUSH, REHYDRATE, PAUSE, PERSIST, PURGE, REGISTER],
-      },
-    }),
-});
-
-export const persistor = persistStore(store);
-
-export type RootState = ReturnType<typeof store.getState>;
-type AppDispatch = typeof store.dispatch;
-
-export const useAppDispatch = () => useDispatch<AppDispatch>();
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
